requirements.txt
langchain==0.1.13
langchain-community==0.0.38
jq==1.8.0
sentence-transformers>=2.7.0
faiss-cpu==1.7.4
transformers==4.40.0
torch==2.2.1
pymongo[srv]
certifi
huggingface_hub>=0.23.0
